remove unnecessary file structure
